% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say hello}
\usage{
hello()
}
\value{
A \code{character(1)} with the word \code{hello}.
}
\description{
Just a short function that prints the word \code{hello}.
}
\examples{
hello()
}
